Add `if` statement syntax
Also factored out test code from nom_parser.rs
